Revert "Take minimum of input rows per chunk and event limit"
This reverts commit c86ed8e0c24207dced7448a0bce9f6154fca4fb2.

The Benchmark tests showed that this did not resolve the file size
difference on the low-end of event numbers. For this reason, I think the
next best explanation is that the fixed minimum file size is given by
the groups necessary to imbue it with the structure we have coded.